+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Current Develop Branch
 
+## 0.3.10 Tue Sep 22 2020
+
+- fix(send): update gas info when sponsorship info loaded
+- fix: sponsor ui into approve page, advacned gas row
+
 ## 0.3.9 Sat Sep 19 2020
 
 - fix: check sponsor info for all tokens (#233)